move netbox models to common
