+++ annet/adapters/netbox/query.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import List
+from typing import List, Union, Iterable, Optional
 
 from annet.storage import Query
 
@@ -9,10 +9,13 @@ class NetboxQuery(Query):
     query: List[str]
 
     @classmethod
-    def new(cls, query, hosts_range) -> "NetboxQuery":
+    def new(
+            cls, query: Union[str, Iterable[str]],
+            hosts_range: Optional[slice] = None,
+    ) -> "NetboxQuery":
         if hosts_range is not None:
             raise ValueError("host_range is not supported")
-        return cls(query=query)
+        return cls(query=list(query))
 
     @property
     def globs(self):
